fix: 清零全部 136 个编译错误，使 analyze/test 通过（CI 可构建 APK/EXE）
- 补齐缺失 import：StarfieldTheme(app_theme)、appThemeProvider(theme_provider)、
  Entry/Todo/伴生类(database)、Value(drift)、EntryRepository/TodoRepository 等
- 修正 Drift API：withUnique→unique；isBetween 包裹 Constant()；count 查询 ?? 0 解包；
  getOverdueTodos 返回 Future.value(<Todo>[])
- 解决 drift Column 与 Flutter Column 命名冲突（hide Column）
- 修正不存在的图标/颜色：Icons.constellation→auto_awesome、Icons.haze→cloud、
  Colors.white87→withOpacity(0.87)；移除 AppBar 非法 subtitle 参数
- 类型修正：Offset→Alignment、ColorTween→Tween<Color>、int→double、
  dart:ui 前缀(Codec/ImmutableBuffer/PictureRecorder)
- 重写 Shadertoy 风格 .frag 为 Flutter 标准 GLSL ES 3.20（FlutterFragCoord + out fragColor）
- 调整 D2 回归断言：锁定“绝不谎报已同步”而非硬编码 error 态（未连接时 truthfully 为 offline）

本地验证：flutter analyze 0 error；flutter test 12/12 通过；着色器编译通过。

+++ lib/features/starmap/widgets/shader_frosted_glass.dart
@@ -132,7 +132,7 @@ class FrostedGlassPainter extends CustomPainter {
     shader.setFloat(9, tint.red / 255.0);
     shader.setFloat(10, tint.green / 255.0);
     shader.setFloat(11, tint.blue / 255.0);
-    shader.setFloat(12, tint.alpha);
+    shader.setFloat(12, tint.alpha.toDouble());
     shader.setFloat(13, borderBrightness);
 
     canvas.drawRect(
